Reverse tmp folder logic so config source can be reused.
